pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "openai>=1.63.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
]
